+++ GameJam/TileEvents/Turret.cs
@@ -30,12 +30,12 @@ namespace GameJam.TileEvents
         {
             Vector2 spawnPosition = new Vector2(renderObject.rectangle.X + gameContext.tileSize * _direction.x, renderObject.rectangle.Y + gameContext.tileSize * _direction.y);
 
-            if (gameContext.room.IsMoveBlocked((int)spawnPosition.x, (int)spawnPosition.y)) return;
+            if (gameContext.room.IsMoveBlocked(spawnPosition.x, spawnPosition.y)) return;
 
             var newBullet = new RenderObject()
             {
                 frames = gameContext.spriteMap.GetPlayerFrames(),
-                rectangle = new Rectangle((int)spawnPosition.x, (int)spawnPosition.y, gameContext.tileSize, gameContext.tileSize),
+                rectangle = new Rectangle(spawnPosition.x, spawnPosition.y, gameContext.tileSize, gameContext.tileSize),
                 objectBehaviour = new Bullet(_direction, 0.5f)
             };
